backend/requirements.txt
dashscope==1.25.6
fastapi==0.128.0
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.40.0
